+++ index.html
@@ -234,7 +234,7 @@
       </section>
     </template>
 
-    <script src="/js/main.js"></script>
+    <script src="/js/main.js" type="module"></script>
     <!--script src="/js/functions.js"></script>-->
 </body>
 </html>
