+++ js/constantsdata.js
@@ -1,0 +1,40 @@
+export const GENERATED_PHOTO_DESCRIPTIONS_COUNT = 25;
+export const AVATAR_COUNT = 6;
+export const MIN_LIKES_COUNT = 15;
+export const MAX_LIKES_COUNT = 200;
+export const MAX_COMMENTS_COUNT = 30;
+export const MAX_COMMENTS_ID = GENERATED_PHOTO_DESCRIPTIONS_COUNT * MAX_COMMENTS_COUNT;
+
+export const DESCRIPTIONS = [
+  'Погожий денек',
+  'День рождения',
+  'Смотрите какая красота',
+  'Угадайте где я?!',
+  'Меня давно просили выложить...',
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export const NAMES = [
+  'Александр',
+  'Алексей',
+  'Али',
+  'Владимир',
+  'Владислав',
+  'Дмитрий',
+  'Дорофей',
+  'Анастасия',
+  'Алёна',
+  'Вика',
+  'Вероника',
+  'Ирина',
+  'Ева',
+  'Кира',
+];
